requirements.txt
langchain-community==0.0.10
langchain==0.1.0
langchain-core==0.1.0
google-cloud-vision>=3.4.0
Pillow>=10.0.0
langchain-google-genai>=0.0.5
google-generativeai>=0.8.5
python-dotenv>=1.0.0
pydantic>=2.0.0
rapidfuzz>=3.0.0
psycopg2-binary>=2.9.0
streamlit
sqlalchemy
redis
pdf2image
fastapi>=0.116.0
uvicorn>=0.35.0
werkzeug>=3.0.0
requests>=2.31.0
email-validator>=2.0.0